:art: Organize code a bit stricter
